requirements.txt
###### Requirements with Version Specifiers ######
numpy==1.21.2
pandas==1.3.3
psycopg2-binary==2.9.1
